% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conway.R
\name{conway}
\alias{conway}
\alias{conway_without_S3}
\title{Main function for Playing Conway's Game of Life}
\usage{
conway(p, r, rule)

conway_without_S3(p, r, rule)
}
\arguments{
\item{p}{Probability as a number between 0 and 1.}

\item{r}{Number of rows.}

\item{rule}{An integer determining the set of rules to be applied in the Game.}
}
\description{
Starts and runs through the Game of Life.
}
\section{Functions}{
\itemize{
\item \code{conway_without_S3}: Function that does not extract information.
}}

\examples{
## example with pattern
decahexagon <- matrix(0, nrow=10, ncol=10)
decahexagon[c(45, 46, 54, 55, 56, 65, 66)] <- 1
decahexagon <- t(decahexagon)
pattern <- matrix(0, nrow=20, ncol=20)
pattern[5:14, 6:15] <- decahexagon
conway(pattern, 20, 2)


## example with probability
conway(0.6, 20, 1)


## Gleiter for ruleset 7
gleiter1 <- matrix(0, nrow=7, ncol=7, byrow=T)
gleiter1[c(10, 12, 22, 24, 26, 28, 30, 34, 45, 47)] <- 1
pattern <- matrix(0, nrow=20, ncol=20)
pattern[7:13, 1:7] <- gleiter1
conway(pattern, 20, 7)


## Gleiter for ruleset 8
gleiter2 <- matrix(0, nrow=7, ncol=7, byrow=T)
gleiter2[c(10, 12, 15, 21, 24, 26, 37, 39, 41)] <- 1
pattern <- matrix(0, nrow=20, ncol=20)
pattern[12:18, 8:14] <- gleiter2
conway(pattern, 20, 8)


}
