% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw}
\alias{draw}
\title{Graphical Display}
\usage{
draw(nw, r)
}
\arguments{
\item{nw}{Network object.}

\item{r}{Number of rows.}
}
\value{
Grid of the actual generation.
}
\description{
Displays a generation graphically on a hexagonal grid.
}
