% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializing.R
\name{initializeNet}
\alias{initializeNet}
\title{Builds network object}
\usage{
initializeNet(p, r)
}
\arguments{
\item{p}{Either a probability p (number between 0 and 1) or
an inputmatrix p with the starting states of the cells (0 or 1).}

\item{r}{Number of rows r.}
}
\value{
List with network object and indexmatrix a.
}
\description{
Initializes object of type network using number of rows r and either probability p or matrix p of starting states of all cells.
}
