% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNodelist.R
\name{neighboursIndices}
\alias{neighboursIndices}
\title{Construction of Edgelists}
\usage{
neighboursIndices(x, a, r)
}
\arguments{
\item{x}{One row of molten indexmatrix with index of row, index of column and index of node.}

\item{a}{Matrix in which entries are indices of nodes.}
}
\value{
A list with two entries: a vector of all indices of 'end 1'-nodes and a vector of all
indices of 'end 2'-nodes (end 1 and end 2 because final network is not directed).
}
\description{
Assigns to every cell the index of its neighbours. And then creates the edgelists that are nessecary for
the representation in a network object.
}
