% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{new.conway}
\alias{new.conway}
\title{Conway Constructor}
\usage{
new.conway(p, r, rule)
}
\arguments{
\item{p}{Probability as a number between 0 and 1.}

\item{r}{Number of rows.}

\item{rule}{An integer determining the set of rules to be applied in the Game.}
}
\description{
Conway Constructor
}
