% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextGeneration.R
\name{nextGeneration}
\alias{nextGeneration}
\title{Updating to Next Generation}
\usage{
nextGeneration(nw = nw, a = a, r = r, rule)
}
\arguments{
\item{nw}{The network object that represents the state of the actual generation.}

\item{a}{Indexmatrix.}

\item{r}{Number of rows.}

\item{rule}{Number of the ruleset that is to be applied.}
}
\value{
The updated network object.
}
\description{
Uses other functions to compute the next generation by respecting the given set of rules and updates
the network object.
}
