% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{numberNeighbours}
\alias{numberNeighbours}
\title{Number of Living Neighbours}
\usage{
numberNeighbours(a, nw = nw)
}
\arguments{
\item{a}{Index of a cell/node.}

\item{nw}{Network object.}
}
\value{
Sum of living neighbours (an integer value).
}
\description{
Computes the sum of all living neighbours of a cell/node.
}
