% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruler.R
\name{ruler}
\alias{ruler}
\title{Ruler}
\usage{
ruler(setruleset = 1, nn, attriMatrix)
}
\arguments{
\item{setruleset}{Number of the set of rules that should be applied to compute the next generation.
The default is Conway's original ruleset 2,3 / 3 (here no. 1).}

\item{nn}{Number of living neighbours of a specific node/cell.}

\item{attriMatrix}{Matrix where entries are attributes of the corresponding node/cell.}
}
\value{
Matrix with the freshly computed attributes of the cells/nodes.
}
\description{
Calls a function for computing the states of the nodes/cells in the next generation corresponding to the indicated ruleset.
}
