% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomStartConfig.R
\name{startconfig}
\alias{startconfig}
\title{Random Starting Configuraion}
\usage{
startconfig(p, r)
}
\arguments{
\item{p}{Probability (number between 0 and 1).}

\item{r}{Number of rows.}
}
\value{
Matrix with entries that are 0 or 1.
}
\description{
Creates a random pattern for the starting generation.
}
